# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaaS Multi-Niche API",
        "description": "API RESTful pour système de réservation multi-niche (paragliding, surfing, diving, etc.)",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server (Développement)"
        },
        {
            "url": "https://api.example.com",
            "description": "API Server (Production)"
        }
    ],
    "paths": {
        "/api/v1/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Liste des activités",
                "description": "Retourne la liste des activités disponibles (paragliding, surfing, etc.)",
                "operationId": "listActivities",
                "parameters": [
                    {
                        "name": "activity_type",
                        "in": "query",
                        "description": "Filtrer par type d'activité",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "paragliding"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des activités",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Activity"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Organisation non trouvée"
                    },
                    "429": {
                        "description": "Rate limit atteint"
                    }
                },
                "security": [
                    {
                        "organization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Activity": {
                "title": "Activity",
                "description": "Modèle d'activité (paragliding, surfing, etc.)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "activity_type": {
                        "type": "string",
                        "example": "paragliding"
                    },
                    "name": {
                        "type": "string",
                        "example": "Vol Tandem Parapente"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "base_price": {
                        "type": "number",
                        "format": "float",
                        "example": 120
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "constraints_config": {
                        "type": "object",
                        "nullable": true
                    },
                    "pricing_config": {
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token. Obtenez le token via /api/v1/auth/login",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "organization": {
                "type": "apiKey",
                "description": "ID de l'organisation (tenant). Requis pour l'isolation multi-tenant.",
                "name": "X-Organization-ID",
                "in": "header"
            }
        }
    },
...
}